+++ src/js/emscripten-settings.ts
@@ -107,13 +107,22 @@ function setEnvironment(env: { [key: string]: string }): PreRunFunc {
  * Mount local directories to the virtual file system. Only for Node.js.
  * @param mounts The list of paths to mount.
  */
-function mountLocalDirectories(mounts: string[]): PreRunFunc {
-  return (Module) => {
-    for (const mount of mounts) {
-      Module.FS.mkdirTree(mount);
-      Module.FS.mount(Module.FS.filesystems.NODEFS, { root: mount }, mount);
-    }
-  };
+function callFsInitHook(
+  fsInit: undefined | ((fs: typeof FS, info: { sitePackages: string }) => void),
+): PreRunFunc[] {
+  if (!fsInit) {
+    return [];
+  }
+  return [
+    async (Module) => {
+      Module.addRunDependency("fsInitHook");
+      try {
+        await fsInit(Module.FS, { sitePackages: Module.API.sitePackages });
+      } finally {
+        Module.removeRunDependency("fsInitHook");
+      }
+    },
+  ];
 }
 
 /**
@@ -131,28 +140,25 @@ function mountLocalDirectories(mounts: string[]): PreRunFunc {
  */
 function installStdlib(stdlibURL: string): PreRunFunc {
   const stdlibPromise: Promise<Uint8Array> = loadBinaryFile(stdlibURL);
-  return (Module) => {
+  return async (Module) => {
     /* @ts-ignore */
     const pymajor = Module._py_version_major();
     /* @ts-ignore */
     const pyminor = Module._py_version_minor();
-
     Module.FS.mkdirTree("/lib");
-    Module.FS.mkdirTree(`/lib/python${pymajor}.${pyminor}/site-packages`);
-
+    Module.API.sitePackages = `/lib/python${pymajor}.${pyminor}/site-packages`;
+    Module.FS.mkdirTree(Module.API.sitePackages);
     Module.addRunDependency("install-stdlib");
 
-    stdlibPromise
-      .then((stdlib: Uint8Array) => {
-        Module.FS.writeFile(`/lib/python${pymajor}${pyminor}.zip`, stdlib);
-      })
-      .catch((e) => {
-        console.error("Error occurred while installing the standard library:");
-        console.error(e);
-      })
-      .finally(() => {
-        Module.removeRunDependency("install-stdlib");
-      });
+    try {
+      const stdlib = await stdlibPromise;
+      Module.FS.writeFile(`/lib/python${pymajor}${pyminor}.zip`, stdlib);
+    } catch (e) {
+      console.error("Error occurred while installing the standard library:");
+      console.error(e);
+    } finally {
+      Module.removeRunDependency("install-stdlib");
+    }
   };
 }
 
@@ -168,11 +174,12 @@ function getFileSystemInitializationFuncs(config: ConfigType): PreRunFunc[] {
     stdLibURL = config.indexURL + "python_stdlib.zip";
   }
 
+  // Note: Hooks are called in **reverse** order of appearance.
   return [
+    ...callFsInitHook(config.fsInit),
     installStdlib(stdLibURL),
     createHomeDirectory(config.env.HOME),
     setEnvironment(config.env),
-    mountLocalDirectories(config._node_mounts),
     initializeNativeFS,
   ];
 }