+++ src/js/pyodide.ts
@@ -48,7 +48,7 @@ export type ConfigType = {
   jsglobals?: object;
   _sysExecutable?: string;
   args: string[];
-  _node_mounts: string[];
+  fsInit?: (FS: FSType, info: { sitePackages: string }) => Promise<void>;
   env: { [key: string]: string };
   packages: string[];
   _makeSnapshot: boolean;
@@ -187,14 +187,12 @@ export async function loadPyodide(
      */
     checkAPIVersion?: boolean;
     /**
-     * Used by the cli runner. If we want to detect a virtual environment from
-     * the host file system, it needs to be visible from when `main()` is
-     * called. The directories in this list will be mounted at the same address
-     * into the Emscripten file system so that virtual environments work in the
-     * cli runner.
-     * @ignore
+     * This is a hook that allows modification of the file system before the
+     * main() function is called and the intereter is started. When this is
+     * called, it is guaranteed that there is an empty site-packages directory.
+     * @experimental
      */
-    _node_mounts?: string[];
+    fsInit?: (FS: FSType, info: { sitePackages: string }) => Promise<void>;
     /** @ignore */
     _makeSnapshot?: boolean;
     /** @ignore */
@@ -220,7 +218,6 @@ export async function loadPyodide(
     stdin: globalThis.prompt ? globalThis.prompt : undefined,
     lockFileURL: indexURL + "pyodide-lock.json",
     args: [],
-    _node_mounts: [],
     env: {},
     packageCacheDir: indexURL,
     packages: [],